added couple mods and updates
tried to update burst fins to also allow speed boost while walking, couldn't do it and ended up not bothering any more

tried to add a vehicle item icon for equippable item icons, but got pissed off by the lack of consistency with vehicle upgrade module use

tried to make a warp vehicle module to pair with the player warp module, but due to the above mentioned problems with the vehicle item icon and the vehicle module use it was quickly abandoned

added miniature suit, an actually working mod for once. Will release it once I have a working icon for it, seems to be finished functionality wise though.

added pickupable vehicles update to allow the cyclops to be picked up. Added because of a person's cyclops getting stuck and me being nice

added another working mod, time control suit, same as miniature suit here I'm waiting for new icon for it before releasing it. Also seems mostly finished functionality wise however

wouldn't normally mass commit like this, but too lazy to sort through what I should keep vs delete and starting to get paranoid with having this many files only on my pc and not on github, so unfortunately a mass commit it is